refactor: 로그아웃 로직 리팩토링
